+++ src/operators/expression/type/isNumber.ts
@@ -1,0 +1,21 @@
+/**
+ * Type Expression Operators: https://docs.mongodb.com/manual/reference/operator/aggregation/#type-expression-operators
+ */
+
+import { computeValue, Options } from "../../../core";
+import { AnyVal, isNil, isNumber, isString, RawObject } from "../../../util";
+
+/**
+ * Checks if the specified expression resolves to a numeric value
+ *
+ * @param obj
+ * @param expr
+ */
+export function $isNumber(
+  obj: RawObject,
+  expr: AnyVal,
+  options?: Options
+): boolean | null {
+  const n = computeValue(obj, expr, null, options);
+  return isNumber(n)
+}